update BaseBoss+LeRauxe, and finish Katherine

+++ MyClaw/BaseCharacter.h
@@ -11,6 +11,7 @@ public:
 	virtual void Logic(uint32_t elapsedTime) override = 0;
 	virtual void Draw() override;
 	virtual D2D1_RECT_F GetRect() override = 0;
+	// TODO: change to int16_t ?
 	virtual pair<D2D1_RECT_F, int8_t> GetAttackRect() = 0; // returns the attack rectangle and the damage it cause
 
 	virtual bool isStanding() const;
@@ -25,7 +26,7 @@ protected:
 	map<string, shared_ptr<Animation>> _animations; // map of all animations of this character. format: [name]=animation
 	D2D1_RECT_F _lastAttackRect; // the player/enemy attackRect (that this character get, e.g. officer use it after player hit him)
 	D2D1_RECT_F _saveCurrRect;
-	int8_t _health;
+	int16_t _health;
 	bool _forward; // TODO use `_isMirrored`
 	bool _isAttack;
 };
